Signal and other system upgrades
Added new caching localization wrapper (not done yet)

and added using that which should fix a ton of localization errors, also various small manual fixes for random errors

Fixed a few simple errors to get all of the errors to show up

Text replaced Xform with *, and removed some unnecessary parentheses

also from other parts of the code as that was easier to do automatically and I left the instances of removed parentheses where it didn't cause an unclear precedence warning

Added the new Godot temporary folder to gitignore

Upgrade kinematic mode code

More small fixes before a bigger bulk edit

Updated setup instructions to mention Godot 4 and some style stuff

that will need to be different in 4.0

Mostly signal fixes but various other small fixes as well

More small fixes all over, compiling image changes, and colour picker

need to verify later that the flags that used to be set on the images work correctly

Corrected most of remaining signal connections to compile

advanced CustomDropDown conversion a bit more

+++ src/space_stage/PlanetSystem.cs
@@ -52,9 +52,9 @@ public class PlanetSystem
             // TODO: processing for non-player cities
             if (planet.IsPlayerOwned)
             {
-                planet.ProcessSpace(elapsed, societyData.SocietyResources);
+                planet.ProcessSpace((float)elapsed, societyData.SocietyResources);
 
-                planet.ProcessResearch(elapsed, societyData);
+                planet.ProcessResearch((float)elapsed, societyData);
 
                 storage += planet.TotalStorageSpace;
                 population += planet.Population;
